Enemy AI pathfinding complete
Have developped the pathfinding algorithm for enemy NPCs to move towards
the player. Only issue is that the pathfinding is not perfect, NPCs can
get stuck behind walls if the route it needs to take to the player takes
it further away from the player before it can get closer. I won't make
the algorithm any more complex than this, it will suffice for the sake
of the project.